G lstarsky

+++ src/main/java/com/example/SerminaProject/Model/Tour.java
@@ -5,7 +5,6 @@ import lombok.Data;
 import lombok.NoArgsConstructor;
 import lombok.AllArgsConstructor;
 
-import java.math.BigDecimal;
 import java.util.List;
 
 @Entity
@@ -26,7 +25,7 @@ public class Tour {
     private String moTa;
 
     @Column(name = "gia")
-    private BigDecimal gia;
+    private Double gia;
 
     // Quan hệ 1-Nhiều với bảng Lịch Trình Tour
     // orphanRemoval = true: Nếu xóa Tour, tự động xóa luôn các trạm bên trong
@@ -57,11 +56,11 @@ public class Tour {
         this.moTa = moTa;
     }
 
-    public BigDecimal getGia() {
+    public Double getGia() {
         return gia;
     }
 
-    public void setGia(BigDecimal gia) {
+    public void setGia(Double gia) {
         this.gia = gia;
     }
